+++ Source/Rendering/ShaderLoader.cpp
@@ -966,12 +966,12 @@ void ShaderLoader::LoadSavedShaders() {
         );
 
         const vec2 uvs[6] = vec2[6](
-            vec2(0.0, 1.0), 
-            vec2(0.0, 0.0), 
-            vec2(1.0, 1.0), 
-            vec2(0.0, 0.0), 
-            vec2(1.0, 0.0),
-            vec2(1.0, 1.0)  
+            vec2(0.0, 0.0), // 0
+            vec2(0.0, 1.0), // 1
+            vec2(1.0, 0.0), // 2
+            vec2(0.0, 1.0), // 3
+            vec2(1.0, 1.0), // 4
+            vec2(1.0, 0.0)  // 5
         );
 
         void main() {